requirements.txt
# Core ML Libraries
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0  # YOLOv8
transformers>=4.30.0  # CLIP
Pillow>=10.0.0

# Computer Vision
opencv-python>=4.8.0
numpy>=1.24.0

# Vector Search
faiss-cpu>=1.7.4  # Use faiss-gpu if you have CUDA
# Alternative: qdrant-client>=1.7.0

# Audio Processing (for voice commands - Phase 2)
openai-whisper>=20230314
sounddevice>=0.4.6

# API and Database
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
sqlalchemy>=2.0.0
pydantic>=2.0.0

# Utilities
tqdm>=4.66.0
python-dotenv>=1.0.0
pandas>=2.0.0